Refactors to remove magnet_tools.py file by moving calculate_bfield_phase_error and calculate_trajectory_straightness into field_generator.py. Prepares for module restructure.
